--- lib/app/features/feed_search/views/pages/chat_advanced_search_page/components/chat_advanced_search_users/chat_advanced_search_users.dart
@@ -1,41 +1,0 @@
-// SPDX-License-Identifier: ice License 1.0
-
-import 'package:flutter/material.dart';
-import 'package:flutter_hooks/flutter_hooks.dart';
-import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:ion/app/extensions/extensions.dart';
-import 'package:ion/app/features/feed/views/components/list_separator/list_separator.dart';
-import 'package:ion/app/features/feed/views/components/post_list/post_list_skeleton.dart';
-import 'package:ion/app/features/feed_search/providers/feed_search_users_provider.dart';
-import 'package:ion/app/features/feed_search/views/components/nothing_is_found/nothing_is_found.dart';
-import 'package:ion/app/features/feed_search/views/pages/feed_advanced_search_page/components/feed_advanced_search_users/feed_advanced_search_user_list_item.dart';
-
-class ChatAdvancedSearchUsers extends HookConsumerWidget {
-  const ChatAdvancedSearchUsers({required this.query, super.key});
-
-  final String query;
-
-  @override
-  Widget build(BuildContext context, WidgetRef ref) {
-    useAutomaticKeepAlive();
-
-    final usersSearchResults = ref.watch(feedSearchUsersProvider(query));
-
-    return usersSearchResults.maybeWhen(
-      data: (userIds) {
-        if (userIds == null || userIds.isEmpty) {
-          return NothingIsFound(
-            title: context.i18n.feed_nothing_found,
-          );
-        }
-
-        return ListView.separated(
-          itemCount: userIds.length,
-          itemBuilder: (context, index) => FeedAdvancedSearchUserListItem(userId: userIds[index]),
-          separatorBuilder: (_, __) => FeedListSeparator(),
-        );
-      },
-      orElse: () => const CustomScrollView(slivers: [PostListSkeleton()]),
-    );
-  }
-}
