+++ lib/app/features/search/model/advanced_search_category.dart
