+++ lib/app/features/search/providers/chat_search_history_provider.dart
