+++ lib/app/features/search/providers/feed_search_history_provider.dart
