+++ lib/app/features/search/views/components/advanced_search_navigation/advanced_search_navigation.dart
