+++ lib/app/features/search/views/components/nothing_is_found/nothing_is_found.dart
