+++ lib/app/features/search/views/components/search_history_clear_confirm/search_history_clear_confirm.dart
