+++ lib/app/features/search/views/components/search_history_empty/search_history_empty.dart
